requirements.txt
streamlit>=1.20.0
pandas>=1.5.0
numpy>=1.23.0
plotly>=5.14.0
statsmodels>=0.14.0
scikit-learn>=1.1.0
openpyxl>=3.0.0
pywebview>=4.0.0